requirements.txt
yfinance
openai
python-telegram-bot==13.15
pyyaml
feedparser

